fix: お気に入り機能を廃止しデータセットにプレビュー・ナビゲーション機能を移植 (#117)
- FavoritePage、useFavoritesフック、favorites API（フロント・サーバー）を完全削除
- AppViewからFavoriteToggleButtonを除去
- Sidebar、App.tsxからfavorites関連のviewMode・ハンドラを除去
- DatasetManagerにMarkdownプレビュー機能を追加（デスクトップ: 分割ペイン、モバイル: 全画面）
- DatasetManagerにアプリ/機能/図解へのリンクナビゲーションを追加
- DatasetItemの型にdiagramタイプを追加
- TypeBadgeのFavoriteItem依存をリテラル型に変更

Closes #117
