+++ viewer/src/components/Sidebar.tsx
@@ -17,14 +17,6 @@ interface ViewModeButtonsDef {
 
 const VIEW_MODE_BUTTONS: ViewModeButtonsDef[] = [
   {
-    mode: "favorites",
-    activeColor: "text-pink-400 bg-pink-400/10",
-    inactiveHover: "text-gray-500 hover:text-pink-400 hover:bg-pink-400/10",
-    title: "お気に入り",
-    path: "M21 8.25c0-2.485-2.099-4.5-4.688-4.5-1.935 0-3.597 1.126-4.312 2.733-.715-1.607-2.377-2.733-4.313-2.733C5.1 3.75 3 5.765 3 8.25c0 7.22 9 12 9 12s9-4.78 9-12z",
-    fillWhenActive: true,
-  },
-  {
     mode: "commands",
     activeColor: "text-cyan-400 bg-cyan-400/10",
     inactiveHover: "text-gray-500 hover:text-cyan-400 hover:bg-cyan-400/10",
@@ -108,9 +100,8 @@ interface SidebarProps {
   isMobile: boolean;
   mobileOpen: boolean;
   onMobileClose: () => void;
-  viewMode: "apps" | "datasets" | "favorites" | "commands" | "config" | "queue" | "scheduler";
+  viewMode: "apps" | "datasets" | "commands" | "config" | "queue" | "scheduler";
   onSelectDatasets: () => void;
-  onSelectFavorites: () => void;
   onSelectCommands: () => void;
   onSelectConfig: () => void;
   onSelectQueue: () => void;
@@ -292,7 +283,6 @@ export function Sidebar({
   onMobileClose,
   viewMode,
   onSelectDatasets,
-  onSelectFavorites,
   onSelectCommands,
   onSelectConfig,
   onSelectQueue,
@@ -306,7 +296,6 @@ export function Sidebar({
   const expanded = !collapsed || hovered;
 
   const modeHandlers: Record<string, () => void> = {
-    favorites: onSelectFavorites,
     commands: onSelectCommands,
     datasets: onSelectDatasets,
     queue: onSelectQueue,
